One Man Revolution V0.4.1
+ Added a place to show the current version
+ Renamed 'Refined Materials' to 'Tier 0 Materials' to better match naming convention of future material tiers
+ Added more code comments to make navigating the file easier

I will be making more frequent, but smaller updates from this point due to frequent crashes in Notepad++ that have resulted in loss of code. A change such as V0.4.1 to V0.4.2 indicates that the game is functionally the same, while some minor visible changes or major invisible changes have been made. A change such as V0.4.2 to V0.5.0 indicates that something functional has changed in the game. Partial changes to game code will be left in comments in minor updates, and will be taken out of comments in major updates when the functionality is confirmed.